Update Oversampling.cpp
Fixed int to long in loop

+++ src/Oversampling.cpp
@@ -26,7 +26,7 @@ unsigned long Oversampling::read(int pin)
   // Can Lead To Overflow
   unsigned long TotalADC = 0;
   for (int i=0; i<_Averaging; i++){
-    for (int i=0; i<OversampleCount; i++){
+    for (long i=0; i<OversampleCount; i++){
       TotalADC = TotalADC + analogRead(pin);
     }
   }
